refactor: atk def speed to float

+++ Assets/Inventory/InventoryScript/InventoryManager.cs
@@ -110,9 +110,9 @@ public class InventoryManager : MonoBehaviour
     public static void UpdateItemInfo(
         string itemName,
         string itemDescription,
-        int HP,
-        int ATK,
-        int DEF,
+        float HP,
+        float ATK,
+        float DEF,
         Sprite itemImage,
         int price
     )
